fix: Add cli extra to tox testing dependencies
Since click was moved to optional [cli] extra, the test environment needs
to include it to run CLI tests. Added 'cli' to extras in [testenv:default]
so that both click and rich are installed during testing.

This fixes the CI test failure where CLI modules fail to import due to
missing click dependency.